third_party/blink/web_tests/platform/generic/fast/dom/DOMImplementation/createDocument-with-used-doctype-expected.txt

document.implementation.createDocument with current document's DOCTYPE.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS doc = document.implementation.createDocument(null, null, document.doctype) did not throw exception.
PASS doc.doctype is doctype
PASS doc.firstChild is doctype
PASS document.doctype is null
PASS successfullyParsed is true

TEST COMPLETE






PASS successfullyParsed is true

TEST COMPLETE
